I new it would come to this

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int prime_factor(void)
 {
-	long long i, number;
+	long int i, number;
 
 	number = 612852475143;
 
@@ -32,7 +32,7 @@ int prime_factor(void)
 		max_prie = number;
 	}
 
-	printf("%lld", max_prime);
+	printf("%ld", max_prime);
 
 	return (0);
 }
